requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
PyPDF2
faiss-cpu
huggingface-hub
